prevent JS error in Django's actions.js
The Django-Admin loads a JS file named admin/js/actions.js for all list views. It handles the action button on top. Since django-filer handles its actions itself, that button must be mimicked, otherwise a JS error is thrown, complaining that this button can not be found in the DOM.