fix: crash in the file detail view
In a recent fix, we introduced a regression where we started to access
width and height of a file. However, this will only work for images and
not for normal files like PDF, Docx etc.

In order to fix this, we need to guard and check if the file is actually
an image before trying to access the attributes.

- Github Issue: #1394

Authored-by: Vinit Kumar <vinit.kumar@kidskonnect.nl>