fix: incompatibility with djangocms-versioning-filer 1.3 was fixed
When used with djangocms-versioning-filer, the `FileAdminChangeForm` might not
have a file field due to missing permissions.
Since djangocms-versioning-filer 1.3 the constructor of `FileAdminChangeForm`
crashes when opening the change URL of a published image.

This patch fixes this issue by adding a check whether the file field is present
before overriding it.

Signed-off-by: Florian Scherf <mail@florianscherf.de>